+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.17
+* `transit_time_statistics` deprecated in favor of `exit_entry_times`.
+* Added `mean_return_times` function for discrete systems.
+* Added `mean_return_times` function for continuous systems.
+
 # 1.16
 * `takens_best_estimate` now returns three arguments, the estimate plus the upper and lower 95%-confidence limits.